Stop using urlkey from CDX. Optimise data structures
Stop getting `urlkey` from CDX. `urlkey` is used to collapse data, we don't
need it to render the results because we already have the `original`
URL.

We used to generate the target URLs from `http://www.` + `domain` + `path`
from `urlkey`. This has many issues because the `urlkey` path contains limited
information compared to the URL path (e.g. all characters are
lowercase). With this change, we use the `original` URL.

All URLS used to start with `www.` (hardcoded). This is no longer
happening.

We used to create `array_Year` structure like this:
```
array(
    0 => array(url1, 2001, 2002, 2003, ...),
    1 => array(url2, 2001, 2003, 2005, ...),
    ....
    N
)
```
And then convert it like this in the `years` structure
```
array(
    0 => array(2001, url1, url2, url3, ...),
    1 => array(2002, url1, url5, url6, ...),
    ...
    15 => array(2016, url6, url1, url2, ...)
)
```
We now use a more efficient method to produce the same result which is
passed to the `d3` rendering function.