Update dependencies and related config
Replace unmaintained `mocha-webpack` with `mochapack` and update related
config which needs minor changes.

Replace unused `extract-text-webpack-plugin` and `clean-webpack-plugin`.
These were leftovers from when using `webpack` for building. They are
not necessary as webpack is used only for unit tests and `rollup` is
used for building.

Update dependencies.

The build files remain exactly the same, so there is no need to update
pkg version.